Crate cw2

Structs

ContractVersion

Constants

CONTRACT

Functions

get_contract_version

get_contract_version can be use in migrate to read the previous version of this contract

query_contract_info

This will make a raw_query to another contract to determine the current version it claims to be. This should not be trusted, but could be used as a quick filter if the other contract exists and claims to be a cw20-base contract for example. (Note: you usually want to require interfaces not implementations of the contracts you compose with, so be careful of overuse)

set_contract_version

set_contract_version should be used in instantiate to store the original version, and after a successful migrate to update it