Function cw2::set_contract_version

pub fn set_contract_version<T: Into<String>, U: Into<String>>(
    store: &mut dyn Storage,
    name: T,
    version: U
) -> StdResult<()>

set_contract_version should be used in instantiate to store the original version, and after a successful migrate to update it